requirements.txt
torch
torch_geometric
pytorch-lightning
hydra-core
pandas
numpy
scikit-learn
rdkit
requests
matplotlib
seaborn
jupyter
